CORE-922: Payment Protocol Request refactor

+++ WalletKitCore/src/crypto/handlers/btc/BRCryptoSupportBTC.c
@@ -16,7 +16,7 @@
 
 private_extern BRCryptoFeeBasis
 cryptoFeeBasisCreateAsBTC (BRCryptoUnit unit,
-                           uint32_t feePerKB,
+                           uint64_t feePerKB,
                            uint32_t sizeInByte) {
     return cryptoFeeBasisCreate (cryptoAmountCreate (unit,
                                                      CRYPTO_FALSE,
